+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+## 1.18.0.1 - 2021-10-07
+
+### Added
+
+- New tests for additional Conditional Types
+
+### Changed
+
+- Nothing
+
+### Deprecated
+
+- Nothing
+
+### Removed
+
+- Nothing
+
+### Fixed
+- Additional Conditional Types moved 
+
 ## 1.18.0 - 2021-05-31
 
 ### Added